fix Type

+++ src/main/java/com/checkskills/qcm/repository/SubscriptionRepository.java
@@ -16,7 +16,7 @@ public interface SubscriptionRepository extends JpaRepository<Subscription, Long
 
     // récupère le total des crédits dont l'utilisateur dispose
     @Query(value = "SELECT SUM(Plan.credits) - (SELECT SUM(credits_used) FROM Subscription WHERE employer_id = :employer_id) FROM Plan LEFT JOIN Subscription ON Plan.id = Subscription.plan_id WHERE (Subscription.employer_id = :employer_id)", nativeQuery = true)
-    int totalCredit(@Param("employer_id") Long employer_id);
+    Long totalCredit(@Param("employer_id") Long employer_id);
 
     // récupère les "subscriptions" de l'utilisateur sur lesquelles il reste des credits, et les ordonne pour avoir les plans "monthly" en premier
     @Query(value = "SELECT subscription.id, subscription.plan_id, subscription.employer_id, subscription.credits_used, plan.credits as plan_credits FROM checkskills.subscription LEFT JOIN plan ON subscription.plan_id = plan.id WHERE employer_id = :employer_id AND (plan.credits - subscription.credits_used > 0) ORDER BY FIELD(plan.type, 'monthly', 'oneshot'), credits_used DESC", nativeQuery = true)
